# docs/v3/openapi3.json
{
	"components": {
		"schemas": {
...
			"handlers.Stub": {
				"type": "object"
			},
...
			"models.EndpointResponse": {
				"properties": {
					"advanced_signatures": {
						"type": "boolean"
					},
					"authentication": {
						"$ref": "#/components/schemas/datastore.EndpointAuthentication"
					},
					"created_at": {
						"type": "string"
					},
					"deleted_at": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"events": {
						"type": "integer"
					},
					"http_timeout": {
						"type": "integer"
					},
					"name": {
						"type": "string"
					},
					"owner_id": {
						"type": "string"
					},
					"project_id": {
						"type": "string"
					},
					"rate_limit": {
						"type": "integer"
					},
					"rate_limit_duration": {
						"type": "integer"
					},
					"secrets": {
						"items": {
							"$ref": "#/components/schemas/datastore.Secret"
						},
						"type": "array"
					},
					"slack_webhook_url": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/datastore.EndpointStatus"
					},
					"support_email": {
						"type": "string"
					},
					"uid": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"models.PagedResponse": {
				"properties": {
					"content": {},
					"pagination": {
						"$ref": "#/components/schemas/datastore.PaginationData"
					}
				},
				"type": "object"
			},
			"models.PortalLink": {
				"properties": {
					"can_manage_endpoint": {
						"description": "Specify whether endpoint management can be done through the Portal Link UI",
						"type": "boolean"
					},
					"endpoints": {
						"description": "IDs of endpoints in this portal link",
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"name": {
						"description": "Portal Link Name",
						"type": "string"
					},
					"owner_id": {
						"description": "Alternatively specify OwnerID, the portal link will inherit all the endpoints with this owner ID",
						"type": "string"
					}
				},
				"type": "object"
			},
			"models.PortalLinkResponse": {
				"properties": {
					"can_manage_endpoint": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string"
					},
					"deleted_at": {
						"type": "string"
					},
					"endpoint_count": {
						"type": "integer"
					},
					"endpoints": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"endpoints_metadata": {
						"items": {
							"$ref": "#/components/schemas/datastore.Endpoint"
						},
						"type": "array"
					},
					"name": {
						"type": "string"
					},
					"owner_id": {
						"type": "string"
					},
					"project_id": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"uid": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"util.ServerResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"in": "header",
				"name": "Authorization",
				"type": "apiKey"
			}
		}
	},
	"info": {
		"contact": {
			"email": "support@getconvoy.io",
			"name": "Convoy Support",
			"url": "https://getconvoy.io/docs"
		},
		"description": "Convoy is a fast and secure webhooks proxy. This document contains datastore.s API specification.",
		"license": {
			"name": "Mozilla Public License 2.0",
			"url": "https://www.mozilla.org/en-US/MPL/2.0/"
		},
		"termsOfService": "https://getconvoy.io/terms",
		"title": "Convoy API Reference",
		"version": "24.1.4"
	},
	"openapi": "3.0.3",
	"paths": {
...
		"/v1/projects/{projectID}/endpoints": {
			"get": {
				"description": "This endpoint fetches an endpoints",
				"operationId": "GetEndpoints",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"schema": {
							"enum": [
								"next",
								"prev"
							],
							"type": "string"
						},
						"x-enum-varnames": [
							"Next",
							"Prev"
						]
					},
					{
						"description": "A pagination cursor to fetch the next page of a list",
						"in": "query",
						"name": "next_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The owner ID of the endpoint",
						"in": "query",
						"name": "ownerId",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The number of items to return per page",
						"in": "query",
						"name": "perPage",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "A pagination cursor to fetch the previous page of a list",
						"in": "query",
						"name": "prev_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The name of the endpoint",
						"in": "query",
						"name": "q",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort order, values are `ASC` or `DESC`, defaults to `DESC`",
						"in": "query",
						"name": "sort",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"allOf": [
														{
															"$ref": "#/components/schemas/models.PagedResponse"
														},
														{
															"properties": {
																"content": {
																	"items": {
																		"$ref": "#/components/schemas/models.EndpointResponse"
																	},
																	"type": "array"
																}
															},
															"type": "object"
														}
													]
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "List all endpoints",
				"tags": [
					"Endpoints"
				]
			},
			"post": {
				"description": "This endpoint creates an endpoint",
				"operationId": "CreateEndpoint",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/models.CreateEndpoint"
							}
						}
					},
					"description": "Endpoint Details",
					"required": true,
					"x-originalParamName": "endpoint"
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/models.EndpointResponse"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Created"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Create an endpoint",
				"tags": [
					"Endpoints"
				]
			}
		},
...
		"/v1/projects/{projectID}/portal-links": {
			"get": {
				"description": "This endpoint fetches multiple portal links",
				"operationId": "LoadPortalLinksPaged",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"schema": {
							"enum": [
								"next",
								"prev"
							],
							"type": "string"
						},
						"x-enum-varnames": [
							"Next",
							"Prev"
						]
					},
					{
						"description": "A pagination cursor to fetch the next page of a list",
						"in": "query",
						"name": "next_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The owner ID of the endpoint",
						"in": "query",
						"name": "ownerId",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The number of items to return per page",
						"in": "query",
						"name": "perPage",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "A pagination cursor to fetch the previous page of a list",
						"in": "query",
						"name": "prev_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The name of the endpoint",
						"in": "query",
						"name": "q",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort order, values are `ASC` or `DESC`, defaults to `DESC`",
						"in": "query",
						"name": "sort",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"allOf": [
														{
															"$ref": "#/components/schemas/models.PagedResponse"
														},
														{
															"properties": {
																"content": {
																	"items": {
																		"$ref": "#/components/schemas/models.PortalLinkResponse"
																	},
																	"type": "array"
																}
															},
															"type": "object"
														}
													]
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "List all portal links",
				"tags": [
					"Portal Links"
				]
			},
			"post": {
				"description": "This endpoint creates a portal link",
				"operationId": "CreatePortalLink",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/models.PortalLink"
							}
						}
					},
					"description": "Portal Link Details",
					"required": true,
					"x-originalParamName": "portallink"
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/models.PortalLinkResponse"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Created"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Create a portal link",
				"tags": [
					"Portal Links"
				]
			}
		},
		"/v1/projects/{projectID}/portal-links/{portalLinkID}": {
			"get": {
				"description": "This endpoint retrieves a portal link by its id.",
				"operationId": "GetPortalLink",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "portal link id",
						"in": "path",
						"name": "portalLinkID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/models.PortalLinkResponse"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Retrieve a portal link",
				"tags": [
					"Portal Links"
				]
			},
			"put": {
				"description": "This endpoint updates a portal link",
				"operationId": "UpdatePortalLink",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "portal link id",
						"in": "path",
						"name": "portalLinkID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/models.PortalLink"
							}
						}
					},
					"description": "Portal Link Details",
					"required": true,
					"x-originalParamName": "portallink"
				},
				"responses": {
					"202": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/models.PortalLinkResponse"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Accepted"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Update a portal link",
				"tags": [
					"Portal Links"
				]
			}
		},
		"/v1/projects/{projectID}/portal-links/{portalLinkID}/revoke": {
			"put": {
				"description": "This endpoint revokes a portal link",
				"operationId": "RevokePortalLink",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "portal link id",
						"in": "path",
						"name": "portalLinkID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Revoke a portal link",
				"tags": [
					"Portal Links"
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://dashboard.getconvoy.io/api"
		}
	],
...
}